+++ sdks/python/hatchet_sdk/worker/runner/runner.py
@@ -178,7 +178,7 @@ class Runner:
                     )
                 )
 
-                log_with_level = logger.info if should_not_retry else logger.error
+                log_with_level = logger.info if should_not_retry else logger.exception
 
                 log_with_level(
                     f"failed step run: {action.action_id}/{action.step_run_id}\n{exc.serialize()}"
@@ -208,7 +208,7 @@ class Runner:
                     )
                 )
 
-                logger.error(
+                logger.exception(
                     f"failed step run: {action.action_id}/{action.step_run_id}\n{exc.serialize()}"
                 )
 
@@ -241,7 +241,7 @@ class Runner:
                     )
                 )
 
-                logger.error(
+                logger.exception(
                     f"failed step run: {action.action_id}/{action.step_run_id}\n{exc.serialize()}"
                 )
 
@@ -269,7 +269,7 @@ class Runner:
                     )
                 )
 
-                logger.error(
+                logger.exception(
                     f"failed step run: {action.action_id}/{action.step_run_id}\n{exc.serialize()}"
                 )